# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Beatboxd (Letterboxd for DJ sets). Generated from\nserver/internal/httpapi handler annotations via swag - run\n`make openapi` in server/ to regenerate after changing a handler.",
        "title": "Beatboxd API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/djs": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "djs"
                ],
                "summary": "Create a DJ (or return the existing one, deduped by spotifyId then slug)",
                "parameters": [
                    {
                        "description": "DJ to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createDjRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Dj"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/djs/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "djs"
                ],
                "summary": "Search DJs by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Dj"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/djs/spotify-search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "djs"
                ],
                "summary": "Search Spotify for artists (fails open to [] on Spotify errors)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SpotifyArtist"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/djs/{id}/reviews": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "djs"
                ],
                "summary": "List reviews for a DJ, cursor-paginated by seenAt desc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DJ id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pagination cursor (RFC3339 seenAt of the last item)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size, 1-50, default 20",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedReviews"
                        }
                    }
                }
            }
        },
        "/api/djs/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "djs"
                ],
                "summary": "Get a DJ by slug, with rating aggregate and recent reviews",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DJ slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DjDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/events": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create an event (or return the existing one, deduped by exact name+venue+date)",
                "parameters": [
                    {
                        "description": "event to create",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/events/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Search events by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Event"
                            }
                        }
                    }
                }
            }
        },
        "/api/events/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get an event by id, with all its reviews",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/EventDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "CrowdVibe": {
            "type": "string",
            "enum": [
                "electric",
                "good",
                "average",
                "dead"
            ],
            "x-enum-varnames": [
                "CrowdVibeElectric",
                "CrowdVibeGood",
                "CrowdVibeAverage",
                "CrowdVibeDead"
            ]
        },
        "Dj": {
            "type": "object",
            "required": [
                "bio",
                "createdAt",
                "createdByUserId",
                "genres",
                "id",
                "imageUrl",
                "name",
                "slug",
                "spotifyId",
                "updatedAt"
            ],
            "properties": {
                "bio": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdByUserId": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "spotifyId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "DjDetailResponse": {
            "type": "object",
            "required": [
                "avgRating",
                "dj",
                "recentReviews",
                "reviewCount"
            ],
            "properties": {
                "avgRating": {
                    "type": "number"
                },
                "dj": {
                    "$ref": "#/definitions/Dj"
                },
                "recentReviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ReviewDTO"
                    }
                },
                "reviewCount": {
                    "type": "integer"
                }
            }
        },
        "Event": {
            "type": "object",
            "required": [
                "city",
                "createdAt",
                "createdByUserId",
                "description",
                "eventDate",
                "id",
                "name",
                "venue"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdByUserId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
        "EventDetailResponse": {
            "type": "object",
            "required": [
                "event",
                "reviews"
            ],
            "properties": {
                "event": {
                    "$ref": "#/definitions/Event"
                },
                "reviews": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ReviewDTO"
                    }
                }
            }
        },
...
        "PaginatedReviews": {
            "type": "object",
            "required": [
                "items",
                "nextCursor"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ReviewDTO"
                    }
                },
                "nextCursor": {
                    "type": "string"
                }
            }
        },
...
        "ReviewDTO": {
            "type": "object",
            "required": [
                "createdAt",
                "crowdVibe",
                "crowdVibeNote",
                "djId",
                "eventId",
                "id",
                "rating",
                "reviewText",
                "seenAt",
                "taggedUsers",
                "updatedAt",
                "userId"
            ],
            "properties": {
                "commentCount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "crowdVibe": {
                    "$ref": "#/definitions/CrowdVibe"
                },
                "crowdVibeNote": {
                    "type": "string"
                },
                "dj": {
                    "$ref": "#/definitions/Dj"
                },
                "djId": {
                    "type": "string"
                },
                "event": {
                    "$ref": "#/definitions/Event"
                },
                "eventId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isLikedByMe": {
                    "type": "boolean"
                },
                "isPopular": {
                    "type": "boolean"
                },
                "likeCount": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "reviewText": {
                    "type": "string"
                },
                "seenAt": {
                    "type": "string"
                },
                "taggedUsers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/User"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "SpotifyArtist": {
            "type": "object",
            "required": [
                "genres",
                "imageUrl",
                "name",
                "spotifyId"
            ],
            "properties": {
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "spotifyId": {
                    "type": "string"
                }
            }
        },
...
        "User": {
            "type": "object",
            "required": [
                "avatarUrl",
                "bio",
                "createdAt",
                "displayName",
                "id",
                "updatedAt",
                "username"
            ],
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "apiError": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "createDjRequest": {
            "type": "object",
            "required": [
                "bio",
                "genres",
                "imageUrl",
                "name",
                "spotifyId"
            ],
            "properties": {
                "bio": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "imageUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "spotifyId": {
                    "type": "string"
                }
            }
        },
        "createEventRequest": {
            "type": "object",
            "required": [
                "city",
                "description",
                "eventDate",
                "name",
                "venue"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
...
        "errorEnvelope": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "$ref": "#/definitions/apiError"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Clerk-issued session JWT, e.g. \"Bearer eyJhbGc...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}